Moved request error shared example to prevent warning during test run
